Input Event Handling Now Works With Multiple Scenes]
Previously, I did not realize that `SDL_Event` must only be polled and/or
pumped in the thread that it was initalized in. Otherwise,
`SDL_PollEvents` consumes all of the events and thus, any events or
functions that are further down the call stack that rely on an
`SDL_Event` will not work.

This was resolved by declaring a static vector<SDL_Event>, which can be
globally retrieved the function `GetFrameEvents()` by any callers that
require an event. As long as these event functions are called within the
game's main loop, everything should work.

Next commit: Figure out how to handle overlapping scenes, right now,
mouse events just work across all of them.